WIP - QAT, util for inference latency and switched from double tensors to float

+++ src/examples/train_script.py
@@ -26,7 +26,7 @@ def train_signs():
     device = "cuda:0" if torch.cuda.is_available() else "cpu"
     print(f"Use {device} for training")
     model.to(device)
-    model.double()
+    model.float()
 
     criterion = nn.CrossEntropyLoss()
